[FIX-#740] Fixed bug in IDPosteriorErrorProbability.cpp
Peptide with score 0 (e-value 10^99) will be ignored in the model testing and fitting. If the score is in the method "get_score_" 0 the method will return NaN. After the method was called, the returned score will be compared to NaN. If it is so, the score will be ignored in the model and set to 1.

Committer: Petra Gutenbrunner <pg12@sanger.ac.uk>