[FIX] enable loading of pepXML files that contain analysis summaries
Some files could not be loaded before due to a segmentation fault.
The problem was that an element named "search_summary" can occur within "analysis_summary" and within "msms_run_summary" (where we parse it), but with different meanings.
Now the complete "analysis_summary" element is skipped to avoid this problem.