[FEATURE,API] Support to wrap MetaInfoInterface in QVariant
These changes allow the MetaInfoInterface metadata class to be converted to
QVariant, and saved using the QVariant API. The changes in this class
would probably never directly be used, but instead in MetaInfoInterface
subclasses to simplify their changes to wrap in QVariant.

Specifically, the Q_DECLARE_METATYPE was used to wrap MetaInfoInterface and the
two required stream operators (using QDataStream) were added.

Note: the stream operators are a bit more kludgy than ideal for
MetaInfoInterface in order to handle DataValue types (i.e. converting
between QString/String to the appropriate DataValue), especially for
StringList, IntList, and DoubleList types.