[FEATURE,API] Add PeakInvestigator as an external library
As suggested in pull request #1007, the PeakInvestigator class has been
moved into an external library as was done for OpenSwath and SuperHirn,
and split into several classes. Specifically, the following supporting
class have been created:

* TarFile for bundling/extracting scans from a MSExperiment to a gzip'd
  tarball.
* AbstractSecureFileTransfer that implements the core functionality for
  the (currently) two classes for SFTP transfers.
* LibSSH2SecureFileTransfer for SFTP transfers using libssh2.
* PSCPSecureFileTransfer for SFTP transfters using PuTTY's scp program.

This commit also adds the CMake files for bulding the library, but does
not touch the OpenMS files.