[FIX] allow user to turn off mass trace length check
The check for the length of the masstrace does not make sense if the
mass trace detection is followed by elution peak detection as in the
MassTraceExtractor where long mass traces are split up. Instead what
happens is that MTD is extremely slow since it finds the same masstraces
again and again but is prevented from adding them to the result and move
on.

This commit allows to set max_trace_length to -1 and skip the length
check.