[FIX] catch end-of-array error
ensure that a RT passed the end of the experiment results in an empty
vector being returned

ensure that if RTBegin returns the very last spectrum, there is no
invalid access to memory through spectrum->getRT().
instead, check end condition *first* in the while loop so that an
end-of-array spectrum access is checked first and the second condition
is not executed at all.