[FIX] PhosphoScoring
To count the matches between theoretical spectrum and experimental spectrum, the look up direction changed. Now each peak of the theoretical spectrum is matched with the peaks in the window_top10. This ensures that for each theoretical peak a possible match is counted.

modified:   ../src/openms/source/ANALYSIS/ID/AScore.cpp