[FEATURE,NOP] filter peptide hits based on protein digestion products
This feature filters peptide hits and it is used as part of the
TOPP IDFilter pipeline.

It reads an input FASTA file and it loads individual FASTAEntry
structures.

Then it matches the accessions of the peptide hit and checks if it is a
valid digestion product of the protein sequence.

If peptide hit do not have a matching accession they are not filtered.

IMPORTANT:This feature is not handling peptide hits with unknown
parameters like UNKNOWN_POSITION or N_TERMINAL_POSITION.
This may cause unexpected behavior.