[FEATURE, PARAM] addPeaks_ is now protected
addPrecursorPeaks and adAbundantImmoniumIons as well.
Instead of one maximal charge parameter, the getSpectrum function now has a min_charge and a max_charge parameter.
A new flag parameter was added to the TSG class. Precursor ions for all charges in the given range can be generated with getSpectrum if necessary.