[FIX] TOPPView IDs from theo spectra
Because of synchronization between peak labels and PeptideHits,
PeakAnnotations were generated after matching to Theoretical Spectra
and these were kept and used after the second time an ID was chosen.
This lead to a changes in the ID view depending on whether you look at an ID for the first time.
Now the labels are synchronized and the theoretical spectrum deleted immediately
and the PeakAnnotations stored in the PeptideHits used right away.
This leads to the sequence coverage being shown immediately and a more consistent representation of IDs.