[FEATURE] OPXLLF: reduced memory usage by moving the parallel loop
previously the parallelized loop was looping over the spectra.
The list of candidates and results had to be enumerated and computed for each spectrum
and this increased the memory usage proportional to the number of used threads.
Now the parallelized loop loops over the candidates and results for one spectrum.

This leaves some code that was parallelized previously out of the parallelization for now, but does not slow down the algorithm significantly, while reducing memory usage by a factor almost equal to the number of used threads when it runs on more than one thread.