[FIX] OPXL: fixed handling of "target+decoy" classification
"target+decoy" peptides were treated as decoys for XLMS data. They are now treated as targets.
Also the reporting of the XLMS specific "xl_target_decoy" MetaValue is now more complete and consistent.
It is reported on both the alpha and beta peptide, but combines the target_decoy state of both peptides: if any of the two peptides was only matched to decoy proteins, the whole cross-link is treated as a decoy. Otherwise it is a target.
This makes e.g. filtering out decoys after FDR calculation more straightforward.