[PARAM, IO] OPXL: changed order and descriptions of output files
The order and descriptions are more in line with XFDR now and the new order reflects the likely usage of the output formats much better with idXML at the top. This will also make it more obvious that idXML is the best format to use in KNIME workflows and it should be the first output of the node.