initial implementation of output prefix parameters
as discussed in https://github.com/OpenMS/OpenMS/issues/4404

implementation for SeedListGenerator currently only works
for consensusXML input. For this tool one could think about
to have the -out and the -out_prefix that could be used
exclusively. for the -out_prefix parameter a question
would be how the names should be generated for the other
input types.