[FEATURE,TEST] Change TextExporter output to display accession details
[FEATURE,TEST] Change TextExporter output to display accession details

If a peptide has multiple protein accessions, the text export now contains multiple rows of the same peptide that are identical except for the accession column. That column contains the corresponding protein name and the start and end position of the match. This change reveals whether a peptide matches multiple times in the same protein, and where. These changes are reflected in outputs of TextExporter tests 3,4 and 5.

Closes #4668