Change TextExporter Output to Display Accession Details (#5137)
* [FEATURE,TEST] Change TextExporter output to display accession details

[FEATURE,TEST] Change TextExporter output to display accession details

If a peptide has multiple protein accessions, the text export now contains multiple rows of the same peptide that are identical except for the accession column. That column contains the corresponding protein name and the start and end position of the match. This change reveals whether a peptide matches multiple times in the same protein, and where. These changes are reflected in outputs of TextExporter tests 3,4 and 5.

Closes #4668

* [FEATURE,TEST] Show accession details in one line

[FEATURE,TEST] Show accession details in one line

The text export contains additional columns start and end. Multiple accessions (also in the same protein) are shown as in this example:

accession | start | end
PROT1;PROT1;PROT2 | 1;2;3 | 14;15;16

These changes are reflected is TextExporter tests 1,3,5,7 and 9

Closes #4668

* Update AUTHORS

* don't print semicolons if no information given