[refactor, style] Use preexisting functions instead of QProcess directly in TVToolDiscovery::getParamFromIni_
Previously TVToolDiscovery::getParamFromIni_ rewrote alot of code that was already existing in the form of ExternalProcess::run.
Now that function is used instead. Additionaly in order to prevent opening to many file handles a new enum was introduced to ExternalProcess to allow an additional parameter to run so that the caller can specify the open mode of started process.

Furthermore the use of TemporaryFiles_ was made thread-safe as well by adding a class mutex protecting the write access to filesnames_.