[fix] Param::hasSection checks whether key has a trailing colon and removes it if so
Depending on whether a section was created with Param::addSection("Section1:Section2") or Param::addSection("Section1:Section2:"), i.e. with or without trailing colon, Param::hasSection would not work properly.
If a section was added Param::addSection("Section1:Section2") and was checked with Param::hasSection("Section1:Section2:") false was returned. Now the trailing in hasSection will be cut off so that it'll always find the right section if it exists.