[fix] Fixed error when running doxygen
Doxygen calls TOPPViewBase while generating the docs without the TOPPViewBase::TOOL_SCAN::SKIP_SCAN set so that TVToolDiscovery is called and tries to generate params for all tools.
Because doxygen is called from somewhere else than "~/build/bin" File::getSiblingExecutable and File::getExecutablePath fail because they mistake the directory doxygen is in for the executable directory.
Therefore said flag for TOPPViewBase needs to be set to prevent that.