[Feature] Adding TOPP tool/util params to .TOPPView.ini (#5256)
* Initial commit

* [perf, bug] Fixed issue with thread-safety, now ussing unordered_map instead of vector

- File::getUniqueName previously used to increment a static counter which caused issues when multithreading i.e. it was not thread-safe
- Now saving param for each tool in an unordered_map (hash table) where the name is the key and the param the value.

* [ref, docs] Added documentation and some refactoring

* [ref] Using function local static variables instead of static class variable

* [feat] ToolsDialog now showing tools and utils. Params are passed to the constructor by TOPPViewBase

* [fix] Make sure file handles / QProcesses are closed properly

* [ref] Using std::string instead of OpenMS::String

* [style]

* [feat] Now adding TOPP tool/util params to TOPPView.ini file

By storing them in the TOPPView.ini file they don't have to be generatered at every start - only when there are changes

* [ref] Removed unnecessary static bool from File::getExecutablePath

* [feat] Added function to check whether a section exists in a param

* [fix] Fixed issue where to many file handles were opened

* [refactor] Now using Param::hasSection in TOPPViewBase

* [test] Added test for new Param::hasSection function

* [fix, test, feat] Added enum to TOPPViewBase to specifier whether TVToolDiscovery should be called

Currently TVToolDiscovery is called to generate params for each tool/util if TOPPView's ini file does not contain them. This was an issue when testing because TVToolDiscovery does not work correctly in that case.

* [fix, docs] Removed static members from TVToolDiscovery and added more documentation

* [refactor, docs] Additional documentation and different variable name

* [refactor]

* [refactor]

* [refactor, fix]

* [fix] Using emplace instead of insert for adding new elements to unordered_map in TVToolDiscovery

* [fix]

* [refactor]

* [refactor, style] Use preexisting functions instead of QProcess directly in TVToolDiscovery::getParamFromIni_

Previously TVToolDiscovery::getParamFromIni_ rewrote alot of code that was already existing in the form of ExternalProcess::run.
Now that function is used instead. Additionaly in order to prevent opening to many file handles a new enum was introduced to ExternalProcess to allow an additional parameter to run so that the caller can specify the open mode of started process.

Furthermore the use of TemporaryFiles_ was made thread-safe as well by adding a class mutex protecting the write access to filesnames_.

* [fix] Prevent unneccessary message to std::err which caused a build failure

* [fix] Param::hasSection checks whether key has a trailing colon and removes it if so

Depending on whether a section was created with Param::addSection("Section1:Section2") or Param::addSection("Section1:Section2:"), i.e. with or without trailing colon, Param::hasSection would not work properly.
If a section was added Param::addSection("Section1:Section2") and was checked with Param::hasSection("Section1:Section2:") false was returned. Now the trailing in hasSection will be cut off so that it'll always find the right section if it exists.

* [fix] Fixed error when running doxygen

Doxygen calls TOPPViewBase while generating the docs without the TOPPViewBase::TOOL_SCAN::SKIP_SCAN set so that TVToolDiscovery is called and tries to generate params for all tools.
Because doxygen is called from somewhere else than "~/build/bin" File::getSiblingExecutable and File::getExecutablePath fail because they mistake the directory doxygen is in for the executable directory.
Therefore said flag for TOPPViewBase needs to be set to prevent that.

Co-authored-by: Timo Sachsenberg <sachsenb@informatik.uni-tuebingen.de>