[FEATURE] New Deisotoping Algorithm (#5306)
* [FEATURE] New Deisotoping Algorithm

[FEATURE] New Deisotoping Algorithm

This implementation is in parts based on Guo Ci Teo et al, DOI: 10.1021/acs.jproteome.0c00544.  It applied a more sophisticated model and considers more combinations as possible isotope peaks. It thus should improve the already existing algorithm, although benchmarking is to be done.

 Added the function Deisotoper::deisotopeAndSingleCharge_exp(...)

* Finished tests

Output of test with piece of an actual spectrum is now compared to a previously created and hand-checked file.

* Finished Algorithm and Tests

* Function documentation

* Update Benchmarking

* Fixed rem_low_intensity functionality, changed testest spectrum and altered order of parameters

* Fixed some corner cases of rem_low_intensities

* Minor style fix

* [NOP] blank lines

* remove unnecessary sorting

* Shorten rem_low_intensity implementation

By using classes ThresholdMower and NLargest

* Adapt tests to ThresholdMower

ThresholdMower also removes peaks with intensity slightly above 0, which should not make a difference for the usage case, but the first tests work with distributions and thus have a lot lower intensity.

* Fixed theoretical spectrum generation

* Sped up spectra_generation experiment

* added evaluation section

* New Benchmarking Code

* Optimize code

* Very optimized

* Approximation of Isotope Distribution

Speeds up the algorithm by a factor of 10.
Approximation from Bellew et al https://dx.doi.org/10.1093/bioinformatics/btl276

* Ready to merge

* Update test results

The differences in the results from the test file constist of two debatable cases. Probably, loosening the KL threshold would lead to the old result again. Maybe requires testing.

* Further optimization of approximation

The e-term is not relevant for the relative height of peaks in one cluster, so it is removed. Also optimized calculation of weight^k .

* remove lambda from approximation

With removal of the e-term, lambda is no longer required to be passed

* improved docu + precalculate more of approximation

* Formatting, small changes

_approximateDistribution is now protected;

tolerance_dalton is independent of q, so it is calculated before that loop;

mono_iso_peak_intensity now only holds a value for monoisotopic peaks and its contents are copied only to those. Improves readability imo;

* added python bindings

* Sorted includes

* Changed filtering parameters

Also added if (make_single_charged) for the sorting, as that should be the only case where sorting is necessary.

* removed confusing num_considered_peaks

* flattened "clusters"-vector for efficiency

* Removed protectes static method

Not too familiar with the C++ subtleties, so I hope this is correct (at least it works). As soon as I include this approximation in
CoarseIsotopePatternGenerator it will change again anyway.

* Revert changes to old method

Not 100% sure if my changes were correct and why I introduced them in the first place.

* Update Python bindings

Co-authored-by: Timo Sachsenberg <sachsenb@informatik.uni-tuebingen.de>