initial implementation of output prefix parameters (#4527)
* initial implementation of output prefix parameters

as discussed in https://github.com/OpenMS/OpenMS/issues/4404

implementation for SeedListGenerator currently only works
for consensusXML input. For this tool one could think about
to have the -out and the -out_prefix that could be used
exclusively. for the -out_prefix parameter a question
would be how the names should be generated for the other
input types.

* Merge remote-tracking branch 'upstream/develop' into feature/float_arr

* fix parameter type

* Update TOPPBase.cpp

* adapt file names

* Update TOPPBase.h

* Update TOPPBase.h

* Apply suggestions from code review

Co-authored-by: Julianus Pfeuffer <pfeuffer@informatik.uni-tuebingen.de>

* export to CTD

* fix some tests

* Update ParamCTDFile.cpp

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>
Co-authored-by: Timo Sachsenberg <sachsenb@informatik.uni-tuebingen.de>
Co-authored-by: Julianus Pfeuffer <pfeuffer@informatik.uni-tuebingen.de>