[pyopenms] fixes some of the newly introduced bugs (#5997)
* Add support for PASEF scans

Previously when processing diaPASEF data with overlapping m/z SWATHs
with different IM this would have to split up into multiple .mzML files
and processed seperately.

These changes enable PASEF with overlapping schemes (e.g. py3) to be
processed as one file.

* add name to authors list

* Ensure proper SWATH window is selected

Similar to PRM, transitions can be found in multiple SWATH windows when
PASEF is performed because there are windows with the same m/z however
different IM. Ensure that extraction occurs in the best SWATH. Best
SWATH is determined by the SWATH in which the transitions's IM is the most
centered.

* Add check for m/z swath

fix bug in which transitions were being extracted in m/z windows that
they were not in.

* clean up empty spaces and unnecessary prints

* add test for PASEF swath window processing

create a test with 3 precursors, over two swath windows with same MZ and
different IM. Test that this can be properly distiguished into 2 SWATH
windows and that each precursor is extracted from the window in which
its IM is the most centered.

* [FIX] create script outputs file # actually used

fixed bug where previously sample file outputs OpenSwath_18 and now
outputs OpenSwath_23 which is the .mzML that this script creates.

* More descriptive debug message

report the IM information on the SWATH window in which the precursor
will be added to.

* If overlapping windows add -sonar/-pasef reminder

If windows are overlapping in m/z or IM it may be because the user
forgot to add the -sonar or -pasef flag so add this reminder.

* [FIX] edit cmake files and outputs so tests pass

previous the tests I added did not pass, now everything should pass

* [FIX] fix test 218 SwathFileConsumer_test.cpp

test was failing because added ion mobility information to SwathMap.h.
Now test passes if default IM at -1.

* add swath file consumer test with IM

* [NOP] some code cleanup

* update CHANGELOG

* [MINOR] debug message fix

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>

* [MINOR] debug message fix

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>

* [MINOR] fix comments

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>

* [MINOR] fix comments

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>

* [MINOR] fix code style

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>

* [MINOR] fix debug message

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>

* [FIX] bracket placement

* explain diaPASEF implementation in OSW Description

update OpenSwath description to describe how precursor swath selection
is implemented in diaPASEF.

* [MINOR] add to descrpition -pasef flag

add comment how this feature is only used if pasef is enabled.

* Make PASEF AND PRM mutually exclusive

Currently not supported to organize by PASEF and PRM. Possibly in future
add IM labels to PRM loop such that if two windows have the same m/z
then pick the one with the best IM.

* [FIX?] test fail on mac due to sorting issues

possibly the mac test is failing because there is ambiguous sorting
(fragment mzs are the same). Change fragment mzs in mzML file and library and this should resolve
sorting issues.

* change sort to stable_sort

stable_sort may be slower however it is guarenteed to preserve ordering.
This is useful if the fragment ion mz are the same and thus ordering is
ambiguous.
This guarentees chromatograms to be identifical if run on different OS.

* fixes some of the newly introduced bugs

* [FIX] OSW Test 23 fail on development

Fix assertion errors on OSW test 23 by creating new functions with
modified assertions. New functions can deal with only 1 fragment ion entry (the sum
of fragment ions) and the precursor ion.

* Use stable_sort() for sorting extraction coords

If windows have the same m/z window for extraction sort() leads to
ambiguous sorting which might not be preserved between operating
systems. Replace with stable_sort() (for the other
prepare_coordinates()) signature to ensure sorting is consistent between
different OS.

Co-authored-by: Joshua <joshuacharkow@gmail.com>
Co-authored-by: Hannes Roest <hannes.rost@utoronto.ca>
Co-authored-by: Joshua Charkow <47336288+jcharkow@users.noreply.github.com>