OpenSwathDecoyGenerator remove duplicates (#6054)
* change switchKR behaviour with nontryptic peptide

Previously if peptide did not end in K or R switchKR would replace the
last AA with a random AA. Here, the behaviour is changed such that if K
or R is not present, we keep the terminal AA. This new behaviour means
that the same nontryptic peptide sequence will always generate the same
decoy peptide.

* Add documentation on switchKR method

Not sure how to add to documentation without changing the format of the
method. So had to change return type from void to peptide

* Change test ouput for new switchKR method

* [FEATURE] check generated decoys are not targets

Add a check to check that decoy sequences generated do not happen to be
target sequences (this will cause an error in OpenSwathWorkflow). Also
added appropriate tests.

Current method works, but is quite slow with larger libraries.

* Speedup searching for decoys in targets

instead of searching using vector, search using unordered map for decoy
sequences in targets which drastically speeds up runtime

* [FIX] protein reference

this allows file to be converted to .tsv without error

* Revert "change switchKR behaviour with nontryptic peptide"

This reverts commit 79a090fb73bfa26db2dfaa05138300accde802aa.
Furthermore, also change switch KR back to a void function (revert
5321022074118851ca895d68525b29dbb862f2f3 partially)

* Revert "Change test ouput for new switchKR method"

This reverts commit e7b5f5d07cdc8d56a6d432131d211b058d99958b.
Change test back to the previous one because changed switchKR behaviour
back to default.

* fix tests

replace tests output to reflect the new changes (to swtichKR for
openswathdecoyGenerator) and for dropping decoy duplicates (MRMDecoy_test)

* [FEATURE] check that decoys are not duplicated

with switchKR feature can randomly get 2 peptide sequences that are not
tryptic and almost identical generate the same decoy
e.g. PEPTIDA --> DITPEPQ and PEPTIDE --> DITPEPQ (Q is randomly chosen
for the final AA twice)

- Add check to ensure that this decoy sequence is not added twice
- modify openswathdecoy_test5 to test for this function
	- since hard to test the random nature of switchKR, add 2
duplicate target peptides and ensure the decoy is only generated once
	- Also add the same target peptide with a modification, and
ensure that this peptide does get a decoy generated for it (because it
is a different sequence)

* [MINOR] fix formatting

* [FIX] typo

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>

* [FIX] typo

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>

* [FIX] typo

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>

* [MINOR] clean code

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>

* [MINOR] fix code style

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>

* [MINOR] fix indetation

* [NOP] fix indetation

* [NOP] fix spacing and indentation using uncrustify

manually go through uncrustify changes and change spacing and tabbing
accordingly.

* change ordering of include statements

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>