[TEST] commented out OpenPepXLS's spec.xml tests
These tests are unstable due to differences in number to string conversions between versions of boost.
This can not be avoided or whitelisted due to the nature of the format (Base64 encoded strings in multi-line blocks).
Currently we see this difference between the boost versions <= 1.71.0 and >=1.78.0. It seems the boost developers are still working on the relevant code and further changes in future versions are likely.

The spec.xml format is only necessary for compatibility with the xQuest Result Viewer, which is low priority right now.
The tests can be activated again, if and when our number to string conversion becomes less dependent on updated external libraries.