[FEATURE, TEST] added test mode flag to xQuest spec.xml output format (#6117)
* [TEST] commented out OpenPepXLS's spec.xml tests

These tests are unstable due to differences in number to string conversions between versions of boost.
This can not be avoided or whitelisted due to the nature of the format (Base64 encoded strings in multi-line blocks).
Currently we see this difference between the boost versions <= 1.71.0 and >=1.78.0. It seems the boost developers are still working on the relevant code and further changes in future versions are likely.

The spec.xml format is only necessary for compatibility with the xQuest Result Viewer, which is low priority right now.
The tests can be activated again, if and when our number to string conversion becomes less dependent on updated external libraries.

* [FEATURE, TEST] added test mode flag to xQuest spec.xml output format

If in test mode, write out raw string of the spectrum, instead of the Base64 encoding. This allows FuzzyDiff to compare the numbers correctly and not fail due to slight differences in floating point number representation.
Tested successfully with the relevant boost versions, that caused the original issue.

* [FIX, TEST] xquest spec.xml format tests, write out string, not vector

* [PYOPENMS] updated XQuestResultXMLFile.pxd