[TEST] remove ms1 from OSW test 17
test was failing in debug mode because it was being caught by the fact
that there is no IM informaiton in the MS1 spectra. To avoid this, turn
off MS1 mode. This invovles updating the test files.

NOTE: Unsure if this is the best way to deal with this exception.