Check for IM when -pasef flag set (#6735)
* Check for IM when -pasef flag set

Errors are not properly caught when dealing with PASEF files. This commit aims to catch errors where there is no IM but the -pasef flag is set

* [FIX] Move IM checking outside of parallel loop

Throwing an exception not compatible inside for loop in parallel coding. Move the check outside the parallel clause

* Check for IM earlier 

Instead of checking for IM in swathmapmass correction do it when load iRT file at the beginning

* Code cleanup

Minor code cleanup, remove unneeded variables and use auto reference.

---------

Co-authored-by: Timo Sachsenberg <timo.sachsenberg@uni-tuebingen.de>
Co-authored-by: Samuel Wein <sam@samwein.com>