Works around issue in apple silicon
On M1 OpenSwathWrokflow_21 fails because a feature gets tagged as `potentialOutlier`. There don't appear to be any differences in the values for the feature, and it looks like a case of something genuinely being an edge case to me. We can work around this by just ignoring all the `potentialOutlier` lines