docker: set contrib image tag to latest unless
this is a complication of having the contrib dockerfile built in a separate repository. During a release, we build it here and the tag is consistent between the contrib image and the other images we build. For non-release pushes, we have to make sure that the tag we use to pull that image is "latest" or else this action will fail on other branches that may not exist in the contrib repository.