Fix mac runner brew issues
This fixes an issue where brew updates packages already present on the runner after we install packages. This was a problem because it could lead to brew trying to overwrite links to things like /usr/local/bin/2to3 with versions installed as a dependency of updates to preinstalled packages (the azure integration specifically requires python3.11 at the time of writing). When we install python with brew we can specify to clobber everything that might already be in /usr/local/.