[DOC] Be specific about Boost flags and improve formatting
Boost installed from brew requires dynamic linking.  This is mentioned
in the install documentation but without saying which flag to pass to
CMake.  This commit resolves that.

It also fixes the formatting for a code block.  The indentation under
the <li> element needs to remain consistent otherwise the \code
markers are ignored and the entire section is treated as code.