Only warn about OpenMS data path if env variable differs from default (#8306)
Previous behavior resulted in pyOpenMS setting the env variable to its default, and later warning about it being set. The warning does not really make sense if the env variable is set to the same value as the default. The suggested behavior only warns if both values differ and notes that the value from the env variable is used.

Co-authored-by: Samuel Wein <sam@samwein.com>