cmake: Don't mangle cached environment variables
When using CMake/CTest to write out a `CMakeCache.txt` file,
environment variables injected into a string variable, which is then
written to a file via CMake, results in the value of those variables
being mangled.

For example, CMake list variables such as `CMAKE_PREFIX_PATH` have
their semicolon separators removed, resulting in an invalid list and
errors downstream about missing libraries.

I'm not sure it's wise to do this in the first place, but I kept this
cache initialization system and fixed the issue by moving it out into
a shell script.