windows/qt: Don't expose library path twice
Primary change: Only export the Qt path once.  This didn't work in the
past but was hidden by the "mangled CMake variables" bug.  This fixed
that bug.

But, I couldn't help but clean up this section so that items are
properly grouped together (e.g., choco installs).  Like all other
sections, `echo` statements that affect the following steps are put at
the end of the run block.