Get WITH_PARQUET=ON working in CI (#8370)
* cmake: Don't mangle cached environment variables

When using CMake/CTest to write out a `CMakeCache.txt` file,
environment variables injected into a string variable, which is then
written to a file via CMake, results in the value of those variables
being mangled.

For example, CMake list variables such as `CMAKE_PREFIX_PATH` have
their semicolon separators removed, resulting in an invalid list and
errors downstream about missing libraries.

I'm not sure it's wise to do this in the first place, but I kept this
cache initialization system and fixed the issue by moving it out into
a shell script.

* parquet: Re-enable parquet support and refactor the conda/CI interaction

Set `WITH_PARQUET` default value back to `ON`.

This is mostly a refactor to make the Conda steps more universal.
That is, for other operating systems and possibly other libraries that
we might need in the future.

* windows/qt: Don't expose library path twice

Primary change: Only export the Qt path once.  This didn't work in the
past but was hidden by the "mangled CMake variables" bug.  This fixed
that bug.

But, I couldn't help but clean up this section so that items are
properly grouped together (e.g., choco installs).  Like all other
sections, `echo` statements that affect the following steps are put at
the end of the run block.

* cmake: CMake uses semicolons as list separators